move each settings section to its own page
