+++ src/components/navigation.tsx
@@ -66,11 +66,11 @@ export function Navigation({ children }: NavigationProps) {
       href: '/settings', 
       icon: IconSettings,
       links: [
-        { label: 'Application', href: '/settings#application', icon: IconApps },
-        { label: 'Announcer', href: '/settings#announcer', icon: IconVolume },
-        { label: 'Scheduler', href: '/settings#scheduler', icon: IconClock },
-        { label: 'Discord', href: '/settings#discord', icon: IconBrandDiscord },
-        { label: 'UI', href: '/settings#ui', icon: IconPaint }
+        { label: 'Application', href: '/settings/application', icon: IconApps },
+        { label: 'Announcer', href: '/settings/announcer', icon: IconVolume },
+        { label: 'Scheduler', href: '/settings/scheduler', icon: IconClock },
+        { label: 'Discord', href: '/settings/discord', icon: IconBrandDiscord },
+        { label: 'UI', href: '/settings/ui', icon: IconPaint }
       ]
     },
     { label: 'Info', href: '/info', icon: IconInfoCircle },
@@ -181,7 +181,7 @@ export function Navigation({ children }: NavigationProps) {
                   href={item.href}
                   label={item.label}
                   leftSection={<item.icon size="1rem" />}
-                  active={mounted && pathname === item.href}
+                  active={mounted && (pathname === item.href || (item.href === '/settings' && pathname?.startsWith('/settings')))}
                   childrenOffset={0}
                   c="#F5F5F5"
                   styles={{
@@ -222,22 +222,7 @@ export function Navigation({ children }: NavigationProps) {
                     }}
                     onClick={(event) => {
                       event.preventDefault()
-                      if (link.href.includes('#')) {
-                        // Handle hash links for same-page navigation
-                        const [path, hash] = link.href.split('#')
-                        if (pathname === path) {
-                          // Already on the correct page, just scroll to section
-                          const element = document.getElementById(hash)
-                          if (element) {
-                            element.scrollIntoView({ behavior: 'smooth', block: 'start' })
-                          }
-                        } else {
-                          // Navigate to page with hash
-                          router.push(link.href)
-                        }
-                      } else {
-                        router.push(link.href)
-                      }
+                      router.push(link.href)
                       if (opened) toggle() // Close mobile menu after navigation only if open
                     }}
                   />
